Thanks to Stephan Kleine for reporting these.
	* src/omp.c (send_to_client, send_certificate): Use %zu for size_t.

	* src/oxpd.c (read_protocol): Cast precision arg to int.

	* src/lsc_user.c (file_utils_copy_file): Use the right format specifiers.

	* src/tasks_sql.h (escalate_1): Return -1 if to_address is NULL.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-manager@6673 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/oxpd.c
@@ -236,7 +236,11 @@ read_protocol (gnutls_session_t* client_session, int client_socket)
                 }
 
 #if TRACE || LOG
-              logf ("<= client %.*s\n", count, from_client + from_client_end);
+              logf ("<= client %.*s\n",
+                    /* Cast is safe because count is bounded by from_client
+                     * size. */
+                    (int) count,
+                    from_client + from_client_end);
 #if TRACE_TEXT
               if (g_strstr_len (from_client + from_client_end,
                                 count,
@@ -244,7 +248,9 @@ read_protocol (gnutls_session_t* client_session, int client_socket)
                 tracef ("<= client  Input may contain password, suppressed.\n");
               else
                 tracef ("<= client  \"%.*s\"\n",
-                        count,
+                        /* Cast is safe because count is bounded by
+                         * from_client size. */
+                        (int) count,
                         from_client + from_client_end);
 #else
               tracef ("<= client  %i bytes\n", count);
